web/requirements.txt
django
djangorestframework
django-request
django-extensions
psycopg2
black
pylint
pylint-django
scapy
channels_redis
gunicorn
redis